Move AirportRemovalArgs to RemoveAirport.h
