+++ src/OpenLoco/src/GameCommands/Airports/RemoveAirport.h
@@ -1,0 +1,29 @@
+#pragma once
+
+#include "GameCommands/GameCommands.h"
+
+namespace OpenLoco::GameCommands
+{
+    struct AirportRemovalArgs
+    {
+        static constexpr auto command = GameCommand::removeAirport;
+
+        AirportRemovalArgs() = default;
+        explicit AirportRemovalArgs(const registers& regs)
+            : pos(regs.ax, regs.cx, regs.di)
+        {
+        }
+
+        // Note: pos.z must be a floored BigZ
+        World::Pos3 pos;
+
+        explicit operator registers() const
+        {
+            registers regs;
+            regs.ax = pos.x;
+            regs.cx = pos.y;
+            regs.di = pos.z;
+            return regs;
+        }
+    };
+}
