+++ src/OpenLoco/src/GameCommands/GameCommands.h
@@ -182,29 +182,6 @@ namespace OpenLoco::GameCommands
         }
     };
 
-    struct AirportRemovalArgs
-    {
-        static constexpr auto command = GameCommand::removeAirport;
-
-        AirportRemovalArgs() = default;
-        explicit AirportRemovalArgs(const registers& regs)
-            : pos(regs.ax, regs.cx, regs.di)
-        {
-        }
-
-        // Note: pos.z must be a floored BigZ
-        World::Pos3 pos;
-
-        explicit operator registers() const
-        {
-            registers regs;
-            regs.ax = pos.x;
-            regs.cx = pos.y;
-            regs.di = pos.z;
-            return regs;
-        }
-    };
-
     struct VehicleAirPlacementArgs
     {
         static constexpr auto command = GameCommand::vehiclePlaceAir;
