+++ src/OpenLoco/src/Windows/Construction/StationTab.cpp
@@ -2,6 +2,7 @@
 #include "Construction.h"
 #include "Drawing/SoftwareDrawingEngine.h"
 #include "GameCommands/Airports/CreateAirport.h"
+#include "GameCommands/Airports/RemoveAirport.h"
 #include "GameCommands/GameCommands.h"
 #include "GameCommands/Road/CreateRoadStation.h"
 #include "GameCommands/Road/RemoveRoadStation.h"
