+++ src/OpenLoco/src/World/CompanyAi.cpp
@@ -3,6 +3,7 @@
 #include "CompanyManager.h"
 #include "Date.h"
 #include "GameCommands/Airports/CreateAirport.h"
+#include "GameCommands/Airports/RemoveAirport.h"
 #include "GameCommands/Company/BuildCompanyHeadquarters.h"
 #include "GameCommands/GameCommands.h"
 #include "GameCommands/Road/CreateRoadMod.h"
